Fix retreival of data from the contract issue
The issue was that the abi I had was incomplete. It was copied from console log after deploy.js execution. A better solution would be to produce a json file with the required information i.e. abi and address. Or get the ABI from the deployed contract via REMIX IDE or from etherscan