Complete front-end testing
The application is running as expected. The only problem right now is that the state is not updated when a transaction is complete. For example, when a user enters lottery the message at the top does not update the number of lottery players. It is the same with pickWinner transaction. Every time a transaction is sent and there is new data on the contract, React state should be updated to show that